Tests/LibWeb/Text/expected/abspos-relayout-insert-under-body.txt

inserted=70,45,30,20
partial-layouts=0 full-layouts=1
PASS
